requirements.txt
streamlit
paho-mqtt
pillow